addを暫定的に定義（フラグレジスタを書き換えない）
